Basic test pattern support for video output devices
Instead of filling video output buffers with 0s, generate and user a
test pattern.

If the -F argument is given, the test pattern is read from the specified
file. The file size must be at equal to or bigger than the image size.

If the -F argument is not given, a basic test pattern is generated
automatically. Every byte is set to the sum of its X and Y coordinates.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>