yavta: Rework to not use video_splitter
video_splitter has to duplicate the video data for all
active outputs, which has obvious memory bandwidth implications.

Rework to bring the (zero copy) buffers back to the ARM and send
to both encoder and render having acquired the appropriate number
of references to the buffer. Only return the buffer to the ISP
output when all references have been released.