yavta: Correct ARM side splitter code for thread safety.
It is incorrect to acquire a MMAL_BUFFER_HEADER_T multiple
times and send it to several components. The correct thing
to do is to use mmal_buffer_header_replicate to copy the
buffer header's contents into multiple new buffers. Each
replication holds a reference on the original buffer in
a thread safe manner, and the original buffer will be
released once all replicas have been released.