Handle V4L2 allocating fewer than nbufs buffers.
There needs to be a 1 to 1 mapping between V4L2 and MMAL buffers.
The MMAL pool is allocated first with nbufs buffers. V4L2 VIDIOC_REQBUF
is then called requesting nbufs, however it can return fewer buffers.
The pairing code then maps up to the number of buffers actually
allocated by V4L2, potentially leaving some unused buffers.

When the main loop comes around trying to find the MMAL buffer to
pass a dequeued V4L2 buffer with it first of all gets the unpaired
buffers, and dereferences a NULL mmal->userdata pointer.

Should we get a buffer with a NULL userdata, drop it. This leaves
the pool only populated with those buffers that are paired up, and
therefore all the other assumptions hold true.